refactor: move fetch mock file
