+++ tests/Services/API.test.ts
@@ -3,7 +3,7 @@ import { config as globalConfig } from '../setupTests';
 import type { CustomHeaders, Method } from '../../src/Calliope/Concerns/CallsApi';
 import { finish } from '../../src/Support/string';
 import InvalidArgumentException from '../../src/Exceptions/InvalidArgumentException';
-import fetchMock from '../fetch-mock';
+import fetchMock from '../mock/fetch-mock';
 
 const url = finish(String(globalConfig.get('baseEndPoint')), '/') + 'users';
