+++ tests/Services/ApiResponseHandler.test.ts
@@ -1,5 +1,5 @@
 import ApiResponseHandler from '../../src/Services/ApiResponseHandler';
-import fetchMock from '../fetch-mock';
+import fetchMock from '../mock/fetch-mock';
 import User from '../mock/Models/User';
 import type { ApiResponse } from '../../src/Contracts/HandlesApiResponse';
 import { API } from '../../src';
