+++ tests/mock/fetch-mock.ts
@@ -1,6 +1,6 @@
-import Collection from '../src/Support/Collection';
-import type { Method } from '../src';
-import { isObjectLiteral } from '../src';
+import Collection from '../../src/Support/Collection';
+import type { Method } from '../../src';
+import { isObjectLiteral } from '../../src';
 // eslint-disable-next-line @typescript-eslint/no-redeclare
 import { Response } from 'cross-fetch';
